+++ Cargo.toml
@@ -17,6 +17,7 @@ members = [
   "crates/mempool_infra",
   "crates/mempool_node",
   "crates/mempool_p2p",
+  "crates/mempool_p2p_types",
   "crates/mempool_test_utils",
   "crates/mempool_types",
   "crates/native_blockifier",
@@ -196,6 +197,8 @@ starknet_gateway_types = { path = "crates/gateway_types", version = "0.0.0" }
 starknet_mempool = { path = "crates/mempool", version = "0.0.0" }
 starknet_mempool_infra = { path = "crates/mempool_infra", version = "0.0.0" }
 starknet_mempool_node = { path = "crates/mempool_node", version = "0.0.0" }
+starknet_mempool_p2p = { path = "crates/mempool_p2p", version = "0.0.0" }
+starknet_mempool_p2p_types = { path = "crates/mempool_p2p_types", version = "0.0.0" }
 starknet_mempool_types = { path = "crates/mempool_types", version = "0.0.0" }
 starknet_patricia = { path = "crates/starknet_patricia", version = "0.0.0" }
 starknet_sierra_compile = { path = "crates/starknet_sierra_compile", version = "0.0.0" }
