+++ crates/mempool_p2p_types/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "starknet_mempool_p2p"
+name = "starknet_mempool_p2p_types"
 version = "0.0.0"
 edition.workspace = true
 license.workspace = true
@@ -11,5 +11,8 @@ workspace = true
 [dependencies]
 async-trait.workspace = true
 papyrus_network.workspace = true
+papyrus_proc_macros.workspace = true
+serde = { workspace = true, features = ["derive"] }
+starknet_api.workspace = true
 starknet_mempool_infra.workspace = true
-starknet_mempool_p2p_types.workspace = true
+thiserror.workspace = true
