+++ crates/mempool_p2p_types/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod communication;
+pub mod errors;
+pub mod mempool_p2p_types;
