+++ crates/mempool_p2p_types/src/mempool_p2p_types.rs
@@ -1,0 +1,3 @@
+use crate::errors::MempoolP2pSenderError;
+
+pub type MempoolP2pSenderResult<T> = Result<T, MempoolP2pSenderError>;
